+++ pkg/lockfile/types.go
@@ -2,6 +2,7 @@ package lockfile
 
 import "github.com/google/osv-scanner/pkg/models"
 
+// TODO(v2): Remove completely
 // TODO(v2): These fields do not need JSON tags I believe
 type PackageDetails struct {
 	Name        string                     `json:"name"`
